Simplify normalize_rules
The first goal of Lrama was generating parse.c file
same with what Bison generates.
For this goal, it was needed to set symbol number and rule number
with same order with Bison.
However Lrama has a lot of tests now. Then it’s not needed to generate
completely same parse.c file with Bison.