`RuleBuilder#numberize_references` should check LHS and RHS for duplication
Before this commit, LHS is not included into check list of `candidates`.
Therefore `$args` in the code below is valid reference and converted
to the reference to LHS (`args`).
This commit change it to check both of LHS and RHS when resolve the named
reference, so that the code is reported as duplicated reference.

```
args : ...
     | args ',' arg_splat
         {
             $$ = rest_arg_append(p, $args, $arg_splat, &@$);
         }
     ;
```