Add support for `parameterizing-redefined` category
Specify this category in the `-W` or `--warning` option to output a warning about the redefinition of parameterizing rules.

Currently, for example, users have no way of knowing the parameterizing rules provided in the standard library are redefined.
Therefore, by specifying the `parameterizing-redefined` option, users will be able to know parameterizing rules have been redefined.