Add diagram generation feature and integrate Railroad Diagrams
Lrama provides an API for generating HTML syntax diagrams. These visual diagrams are highly useful as grammar development tools and can also serve as a form of automatic self-documentation.

![Syntax Diagrams](https://github.com/user-attachments/assets/5d9bca77-93fd-4416-bc24-9a0f70693a22)

This feature was inspired by the functionality of [Chevrotain](https://chevrotain.io/docs/), a parser construction toolkit for JavaScript.

The main motivation for adding this feature was to make the structure of `parse.y`, and thus the structure of Ruby's grammar, easier to understand. I thought that generating syntax diagrams in a more human-readable format would be beneficial.

These diagrams are highly useful as grammar development tools and can also serve as automatic self-documentation, which I believe is a significant advancement.

Chevrotain used [tabatkins/railroad-diagrams](https://github.com/tabatkins/railroad-diagrams), but this library only supported JavaScript and Python.

To address this, I created a Ruby library and integrated it into Lrama:
[https://github.com/ydah/railroad_diagrams](https://github.com/ydah/railroad_diagrams)

This makes collaboration with Lrama easier and allows improvements to the library based on Lrama's needs. For these reasons, I decided to release it as a Ruby gem.