State#transition: Prefer Enumerable#find
Using `find` is simpler and faster than explicitly iterating with `each`.

The meaning of the code changes slightly; the older finds the last
element, but the newer finds the first one. But the result should be
unchanged because there should be no elements with the same `next_sym`.

This speeds up Ruby's parse.y processing by 5.9% (1.97s -> 1.86s).